chore: restructure did:ebsi plugin folder
