+++ packages/snap/src/did/ebsi/constants.ts
