+++ packages/snap/src/did/ebsi/ebsiDidOnboarding.ts
