+++ packages/snap/src/did/ebsi/ebsiDidResolver.ts
