+++ packages/snap/src/did/ebsi/ebsiDidUtils.ts
