--- packages/snap/src/did/ebsi/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * Provides `did:ebsi` {@link @veramo/did-provider-ebsi#EbsiDIDProvider | identifier provider } for the
- * {@link @veramo/did-manager#DIDManager}
- *
- * @packageDocumentation
- */
-export { EbsiDIDProvider } from './ebsiDidProvider.js';
-export { ebsiDidResolver } from './ebsiDidResolver.js';
-export * from './types/ebsiProviderTypes.js';
