+++ packages/snap/src/did/ebsi/types/ebsiProviderTypes.ts
